Make SolidusAdmin's testing support code require-able
I need to move some of SolidusAdmin's specs into
solidus_legacy_promotions. When doing that, it's necessary to require
these, so it'd be nice if they lived in `lib/`.
