+++ admin/spec/spec_helper.rb
@@ -84,6 +84,8 @@ require "rails/version"
 require "rails/generators"
 require "rails/generators/app_base"
 require "rails/generators/testing/behavior"
+require "solidus_admin/testing_support/component_helpers"
+require "solidus_admin/testing_support/feature_helpers"
 
 # AXE - ACCESSIBILITY
 require 'axe-rspec'
@@ -112,12 +114,12 @@ RSpec.configure do |config|
 
   config.include FactoryBot::Syntax::Methods
 
-  config.include SolidusAdmin::FeatureHelpers, type: :feature
+  config.include SolidusAdmin::TestingSupport::FeatureHelpers, type: :feature
 
   config.include Capybara::RSpecMatchers, type: :component
   config.include ViewComponent::TestHelpers, type: :component
   config.include ViewComponent::SystemTestHelpers, type: :component
-  config.include SolidusAdmin::ComponentHelpers, type: :component
+  config.include SolidusAdmin::TestingSupport::ComponentHelpers, type: :component
 
   config.include Rails::Generators::Testing::Behavior, type: :generator
   config.include FileUtils, type: :generator
